projects/04-automated-data-pipeline/src/requirements.txt
    google-cloud-storage
    google-cloud-bigquery